pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-collector"
version = "1.0.0"
description = "Vietnamese Stock Market Data Collector"
requires-python = ">=3.10"
dependencies = [
    "vnstock>=3.0.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",
    "apscheduler>=3.10",
    "click>=8.1",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "rich>=13.0",
]

[project.scripts]
stock-collector = "stock_collector.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
